+++ src/commands/ping.js
@@ -1,8 +1,10 @@
 module.exports = {
 	name: 'ping',
 	description: 'Ping!',
-	async execute(message, args) {
+	async execute(message, args, roleId) {
+		const guild = message.guild;
+
 		await message.reply('Pong.');
-        return;
+		return;
 	},
 };