+++ src/copy.js
@@ -1,0 +1,5 @@
+module.exports = {
+	ALREADY_HAS_ROLE: 'You already have the \`Member\` role.',
+    ALREADY_CLAIMED: "Nice to see you again. \`Member\` role granted.",
+    PUNT_TO_DM: "We'll need to chat in private. Check your DM to continue."
+};